refactor rest api
