+++ api/v1/endpoints/websocket.py
@@ -1,0 +1,39 @@
+from fastapi import APIRouter, WebSocket, WebSocketDisconnect
+from fastapi.responses import JSONResponse
+from typing import List
+
+
+class ConnectionManager:
+    def __init__(self):
+        self.connections: List[WebSocket] = []
+
+    async def connect(self, websocket: WebSocket):
+        await websocket.accept()
+        self.connections.append(websocket)
+
+    def disconnect(self, websocket: WebSocket):
+        self.connections.remove(websocket)
+
+    async def broadcast(self, data: str):
+        for connection in self.connections:
+            await connection.send_text(data)
+
+
+manager = ConnectionManager()
+Router = APIRouter()
+
+
+@Router.websocket("/ws")
+async def websocket_endpoint(websocket: WebSocket):
+    await manager.connect(websocket)
+    try:
+        while True:
+            data = await websocket.receive_text()  # client 메시지 대기
+            await manager.broadcast(data)  # client에 메시지 전달
+    except WebSocketDisconnect:
+        manager.disconnect(websocket)
+
+
+@Router.get("/")
+def say_hi():
+    return JSONResponse({'message': 'Hi'})
